% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeHR.R
\name{computeHR}
\alias{computeHR}
\title{CardiacDP - computeHR()}
\usage{
computeHR(
  file_path,
  reduce_res = 0.01,
  pop_size = 10L,
  max_gen = 20L,
  patience = 2L,
  an_in = 1,
  acf_thres = 0.5,
  lr_thres = 0.7
)
}
\arguments{
\item{file_path}{Designate the path to your file, must be a .zip or .csv file}

\item{reduce_res}{Time interval of reduced resolution (seconds), by default 0.01}

\item{pop_size}{Number of populations used in the genetic algorithm, by default 10L}

\item{max_gen}{Maximum number of generations in the genetic algorithm, by default 20L}

\item{patience}{Patience threshold (maximum number of generations with no further changes) in the genetic algorithm, by default 2L}

\item{an_in}{Analysis interval (length of a sequence; in minute), by default 1}

\item{acf_thres}{Threshold used in ACF to classify periodic oscillations from aperiodic noises, by default 0.5}

\item{lr_thres}{Linear regression r-sq threshold in extrapolating the tracking index, by default 0.7}
}
\value{
The positions (in indices) and durations of the sub-sequences (finalsubseq) and the corresponding candidate HR (candidateHR) obtained from the genetic algorithm, and the final results evaluating the candidates by autocorrelation values (results_ACF) or the tracking index (results_TI), which contains the details of the subsequences after checking for resolution (subseqHR with Time_min column), the weighted heart rate per sequence (weightedHR with Time_min column) and a plot (plot). Additionally, CSV files (subseqHR and weightedHR) and PNG files (plots) are automatically saved to the current working directory for each channel and method.
}
\description{
Employing the autocorrelation function (ACF) with a genetic algorithm framework to locate periodic sub-sequences within each sequence. From the candidate heart rates of these sub-sequences, the final results are either evaluated based on the autocorrelation value or a tracking index (TI).
}
\examples{
\dontrun{
# use the default parameters to analyse a zip file
# the collatedata function will be called automatically
computeHR(file_path = "data.zip")
}

\dontrun{
# use the default parameters to analyse a csv file
computeHR(file_path = "data.csv")
}

\dontrun{
# use customized parameters to analyse a zip file
computeHR("data.zip", reduce_res = 0.1, max_gen = 30L, lr_thres = 0.8)
}

\dontrun{
# use custom parameters to analyse a csv file
computeHR("data.csv", pop_size = 20L, an_in = 5, acf_thres = 0.6)
}
}
