% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collatedata.R
\name{collatedata}
\alias{collatedata}
\title{CardiacDP - collatedata()}
\usage{
collatedata(file_path)
}
\arguments{
\item{file_path}{Designate the path to your file, must be a .zip file}
}
\value{
A single collated data table
}
\description{
Automatically read and collate separate .csv files in chronological order as inferred by the file names and in hierarchy.
}
\examples{
\dontrun{
collatedata(file_path = "path/to/your/file.zip")
}
}
